Rename adapter for consistency, cleanup export
