+++ addon/index.js
@@ -1,5 +1,7 @@
-import EmberDataUtilsJSONAPIAdapter from 'ember-data-utils/adapters/ember-data-utils-jsonapi';
+import EmberDataUtilsJSONAPIAdapter from 'ember-data-utils/adapters/ember-data-utils-json-api';
+import EmberDataUtilsJSONAPISerializer from 'ember-data-utils/serializers/ember-data-utils-json-api';
 
 export {
-  EmberDataUtilsJSONAPIAdapter
+  EmberDataUtilsJSONAPIAdapter,
+  EmberDataUtilsJSONAPISerializer
 }
