--- app/utils/parse-query-to-function.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-data-utils/utils/parse-query-to-function';
